backend/requirements.txt
# Core libraries for the API
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
pillow


# Smaller, more focused LangChain packages
langchain-core
langchain-openai
langgraph